Removing flannel.exe from /payload/cni WMCO dir. (#2857)
In order to be able to use kubernetes 1.23 CNI plugins
the flannel service had to be removed. This work has
been addressed in:
Ref: https://github.com/openshift/windows-machine-config-operator/pull/841

This patch removes the flannel.exe binary from the
WMCO operator's expected binaries (under /payload/cni)
and at the same time fixes an issue with the output
message, which was always displaying the same /payload
directory.